+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-22T13:05:42.300Z",
+    "2026-04-22T15:10:21.187Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {